+++ tests/minimal_test/MinimalTest.hpp
@@ -82,6 +82,7 @@ class MinimalTest : public CPPUNIT_NS::TestFixture {
 		CPPUNIT_TEST(testAddBondOperator);
 		CPPUNIT_TEST(testRemoveBondOperator);
 		CPPUNIT_TEST(testMutateAtomOperator);
+		CPPUNIT_TEST(testInterlayAtomOperator);
 		CPPUNIT_TEST(testMolpher);
 		CPPUNIT_TEST(testTree);
 		CPPUNIT_TEST(testExplorationData);
@@ -105,6 +106,7 @@ private:
 	void testAddBondOperator();
 	void testRemoveBondOperator();
 	void testMutateAtomOperator();
+	void testInterlayAtomOperator();
 	void testMolpher();
     void testTree();
     void testExplorationData();